Final Adjustments
Removed a lot of unnecessary files. Updated components and design for
responsive design for tablets and mobile. Cleaned up features so that
they make more sense such as the dynamic bar chart having a label,
geolocation services button being removed for now, slight adjustments to
the jssor slider, and removed the admin features for now but will add
them later and fix them.